agent/requirements.txt
livekit-agents[openai,cartesia,turn-detector]~=1.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.0.0
pydantic==2.5.0
